Removed models dir, moved lib -> utils
