--- webapp/src/models/venue.ts
@@ -1,4 +1,0 @@
-export interface Venue {
-  name: string;
-  location: string;
-}
