+++ webapp/src/pages/home/home.tsx
@@ -3,8 +3,8 @@ import { jsx, Box, Flex, Spinner } from 'theme-ui'
 import { useState, useEffect, Dispatch, SetStateAction } from 'react'
 import useSWR from 'swr'
 
-import { Venue } from '../../models/venue'
-import fetcher, { POST } from '../../lib/fetcher'
+import { Venue } from '../../components/Search/Search'
+import fetcher, { POST } from '../../utils/fetcher'
 import Layout from '../../components/Layout/Layout'
 import Search from '../../components/Search/Search'
 import VenueResult from '../../components/VenueResult/VenueResult'
